+++ src/main/java/com/github/wintersteve25/tau/menu/TauContainerScreen.java
@@ -3,9 +3,9 @@ package com.github.wintersteve25.tau.menu;
 import com.github.wintersteve25.tau.build.BuildContext;
 import com.github.wintersteve25.tau.build.UIBuilder;
 import com.github.wintersteve25.tau.components.base.DynamicUIComponent;
+import com.github.wintersteve25.tau.layout.Axis;
 import com.github.wintersteve25.tau.layout.Layout;
 import com.github.wintersteve25.tau.theme.Theme;
-import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Renderable;
 import net.minecraft.client.gui.components.events.GuiEventListener;
@@ -29,8 +29,8 @@ public class TauContainerScreen extends AbstractContainerScreen<TauContainerMenu
 
     private boolean built;
 
-    public TauContainerScreen(TauContainerMenu pMenu, Inventory pPlayerInventory, UIMenu uiMenu, boolean renderBackground, Theme theme) {
-        super(pMenu, pPlayerInventory, Component.empty());
+    public TauContainerScreen(TauContainerMenu pMenu, Inventory pPlayerInventory, UIMenu uiMenu, boolean renderBackground, Theme theme, Component title) {
+        super(pMenu, pPlayerInventory, title);
         this.uiMenu = uiMenu;
         this.renderBackground = renderBackground;
         this.theme = theme;
@@ -44,12 +44,18 @@ public class TauContainerScreen extends AbstractContainerScreen<TauContainerMenu
         Layout layout = new Layout(uiMenu.getSize().x, uiMenu.getSize().y);
         leftPos = uiMenu.getLeftPos(layout, width, height);
         topPos = uiMenu.getTopPos(layout, width, height);
+        
+        layout.pushOffset(Axis.HORIZONTAL, leftPos);
+        layout.pushOffset(Axis.VERTICAL, topPos);
 
         components.clear();
         tooltips.clear();
         dynamicUIComponents.clear();
 
         UIBuilder.build(layout, theme, uiMenu.build(layout, theme, getMenu()), new BuildContext(components, tooltips, dynamicUIComponents, (List<GuiEventListener>) children(), new ArrayList<>()));
+        
+        layout.popOffset(Axis.HORIZONTAL);
+        layout.popOffset(Axis.VERTICAL);
 
         built = true;
     }
@@ -60,15 +66,9 @@ public class TauContainerScreen extends AbstractContainerScreen<TauContainerMenu
             renderTransparentBackground(graphics);
         }
         
-        PoseStack poseStack = graphics.pose();
-        poseStack.pushPose();
-        poseStack.translate(leftPos, topPos, 0);
-
         for (Renderable component : components) {
             component.render(graphics, pMouseX, pMouseY, pPartialTick);
         }
-
-        poseStack.popPose();
     }
 
     @Override
@@ -76,15 +76,9 @@ public class TauContainerScreen extends AbstractContainerScreen<TauContainerMenu
         super.render(guiGraphics, mouseX, mouseY, partialTick);
         renderTooltip(guiGraphics, mouseX, mouseY);
         
-        PoseStack poseStack = guiGraphics.pose();
-        poseStack.pushPose();
-        poseStack.translate(leftPos, topPos, 0);
-        
         for (Renderable tooltip : tooltips) {
             tooltip.render(guiGraphics, mouseX, mouseY, partialTick);
         }
-
-        poseStack.popPose();
     }
 
     @Override