+++ src/main/java/com/github/wintersteve25/tau/utils/AnimatedClientTooltip.java
@@ -1,0 +1,31 @@
+package com.github.wintersteve25.tau.utils;
+
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.network.chat.Component;
+import org.joml.Matrix4f;
+
+public class AnimatedClientTooltip implements ClientTooltipComponent {
+    
+    private Variable<Component> text;
+
+    public AnimatedClientTooltip(Variable<Component> text) {
+        this.text = text;
+    }
+
+    @Override
+    public void renderText(Font font, int mouseX, int mouseY, Matrix4f matrix, MultiBufferSource.BufferSource bufferSource) {
+        font.drawInBatch(this.text.getValue(), (float)mouseX, (float)mouseY, -1, true, matrix, bufferSource, Font.DisplayMode.NORMAL, 0, 15728880);
+    }
+
+    @Override
+    public int getHeight() {
+        return 10;
+    }
+
+    @Override
+    public int getWidth(Font font) {
+        return font.width(this.text.getValue());
+    }
+}